refactor: refactor main-cart and mini-cart to prepare for redux
